Fix driver deletion when --sarif; Improve entry point a little bit

+++ pipeline/index.js
@@ -775,7 +775,8 @@ async function runPipeline(pkgName) {
                     propBlacklist,
                     writeOnDetect: true,
                     recordAllFunCalls: true,
-                    injectForIn: forInRun
+                    injectForIn: forInRun,
+                    repoPath
                 },
                 EXCLUDE_ANALYSIS_KEYWORDS,
                 execFile
@@ -1154,7 +1155,9 @@ async function run() {
     });
 
     // delete tmp driver
-    fs.rmSync(driverDir, {recursive: true, force: true});
+    if (!cliArgs.sarif) {
+        fs.rmSync(driverDir, {recursive: true, force: true});
+    }
 
     closeConnection();
 }
